fix runtime error with no user logged in

+++ src/components/postItem/postItem.jsx
@@ -22,42 +22,12 @@ function PostItem({ id, comment, username }) {
   });
 
   const handleDelete = () => {
-    if (window.confirm('Delete this entry?')) {
-      deleteBurn({ id: id });
+    if (user !== undefined && user.user_metadata.username === username) {
+      if (window.confirm('Delete this entry?')) {
+        deleteBurn({ id: id });
+      }
     }
   };
-
-  // const handleDelete2 = async (event) => {
-  //   const result = confirm('Delete this entry?');
-  //   if (result) {
-  //     // in case we need to track if something is loading for UX
-  //     // setLoading(true);
-
-  //     try {
-  //       let res = await fetch('/getBurns', {
-  //         method: 'DELETE',
-  //         headers: {
-  //           'Content-Type': 'application/json',
-  //         },
-  //         body: JSON.stringify({
-  //           id: id,
-  //         }),
-  //       });
-
-  //       // 204 status "No Content" for delete requests
-  //       if (res.status === 204) {
-  //         setMessage('Error occurred in the delete request');
-  //       }
-  //     } catch (err) {
-  //       console.log(err);
-  //       setMessage(
-  //         "Fetch didn't happen - Error occurred fetching data in delete request"
-  //       );
-  //     }
-
-  //     setLoading(false); // Set loading back to false after the API call is completed
-  //   }
-  // };
 
   // update function
   // modal shows if logged in user matches user
@@ -66,9 +36,10 @@ function PostItem({ id, comment, username }) {
 
     const username = event.target.parentElement.dataset.user;
 
-    if (user.user_metadata.username === username) {
+    if (user !== undefined && user.user_metadata.username === username) {
       setModalState(!modalState);
     }
+    setLoading(false);
   };
 
   return (
